backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/ML dependencies
groq==0.4.1

# Fast image generation - using OpenAI DALL-E mini alternative
openai==1.3.0

# Image processing
pillow==10.0.1
numpy==1.24.3

# HTTP requests
requests==2.31.0
httpx==0.25.0

# Environment and configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Caching
diskcache==5.6.3 